# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cairn",
    "description": "\nCairn is a knowledge base platform and RAG infrastructure server.\n\nAgents are clients: point at this URL, present an API key, send a query and a\nknowledge base ID, and receive ranked, cited passages.\n\n**Phase 0** ships authentication only. Knowledge base, ingestion, and retrieval\nendpoints arrive in Phase 2.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/mcp-service/config": {
      "put": {
        "tags": [
          "mcp management"
        ],
        "summary": "Configure Service",
        "operationId": "configure_service_v1_mcp_service_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPServiceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mcp-service/actions": {
      "post": {
        "tags": [
          "mcp management"
        ],
        "summary": "Service Action",
        "operationId": "service_action_v1_mcp_service_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPServiceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mcp-service/logs": {
      "get": {
        "tags": [
          "mcp management"
        ],
        "summary": "Service Logs",
        "operationId": "service_logs_v1_mcp_service_logs_get",
        "parameters": [
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "INFO",
                    "WARNING",
                    "ERROR"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPLogs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mcp-service/logs/download": {
      "get": {
        "tags": [
          "mcp management"
        ],
        "summary": "Download Logs",
        "operationId": "download_logs_v1_mcp_service_logs_download_get",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "INFO",
                    "WARNING",
                    "ERROR"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/model-providers": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Providers",
        "operationId": "list_providers_v1_model_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderResponse"
                  },
                  "type": "array",
                  "title": "Response List Providers V1 Model Providers Get"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Create Provider",
        "operationId": "create_provider_v1_model_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model-providers/{provider_id}": {
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Delete Provider",
        "operationId": "delete_provider_v1_model_providers__provider_id__delete",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models",
        "operationId": "list_models_v1_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelResponse"
                  },
                  "type": "array",
                  "title": "Response List Models V1 Models Get"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Create Model",
        "operationId": "create_model_v1_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {}
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {}
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models/{model_id}": {
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Delete Model",
        "operationId": "delete_model_v1_models__model_id__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models/{model_id}/test": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Test Model",
        "operationId": "test_model_v1_models__model_id__test_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelTestResponse"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateModelRequest": {
        "properties": {
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "model_key": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Model Key"
          },
          "display_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Display Name"
          },
          "capability": {
            "type": "string",
            "const": "embedding",
            "title": "Capability",
            "default": "embedding"
          },
          "dimension": {
            "type": "integer",
            "maximum": 65536.0,
            "minimum": 1.0,
            "title": "Dimension"
          },
          "max_input_tokens": {
            "type": "integer",
            "maximum": 32768.0,
            "minimum": 1.0,
            "title": "Max Input Tokens"
          },
          "normalize": {
            "type": "boolean",
            "title": "Normalize",
            "default": true
          },
          "query_prefix": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Prefix"
          },
          "optimal_batch_size": {
            "type": "integer",
            "maximum": 1024.0,
            "minimum": 1.0,
            "title": "Optimal Batch Size",
            "default": 16
          },
          "tokenizer_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Tokenizer Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider_id",
          "model_key",
          "display_name",
          "dimension",
          "max_input_tokens",
          "tokenizer_id"
        ],
        "title": "CreateModelRequest"
      },
      "CreateProviderRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "family": {
            "type": "string",
            "enum": [
              "tei",
              "infinity"
            ],
            "title": "Family"
          },
          "base_url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Base Url"
          },
          "config": {
            "$ref": "#/components/schemas/ProviderConfigRequest"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "family",
          "base_url",
          "config"
        ],
        "title": "CreateProviderRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MCPAction": {
        "properties": {
          "expected_generation": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Generation"
          },
          "action": {
            "type": "string",
            "enum": [
              "start",
              "stop",
              "restart"
            ],
            "title": "Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_generation",
          "action"
        ],
        "title": "MCPAction"
      },
      "MCPConfigUpdate": {
        "properties": {
          "expected_generation": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Generation"
          },
          "config": {
            "$ref": "#/components/schemas/ManagedMCPConfig"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_generation",
          "config"
        ],
        "title": "MCPConfigUpdate"
      },
...
      "MCPLogs": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MCPLogView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "MCPLogs"
      },
      "MCPServiceView": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ManagedMCPConfig"
          },
          "generation": {
            "type": "integer",
            "title": "Generation"
          },
          "observed_generation": {
            "type": "integer",
            "title": "Observed Generation"
          },
          "desired_state": {
            "type": "string",
            "enum": [
              "running",
              "stopped"
            ],
            "title": "Desired State"
          },
          "state": {
            "type": "string",
            "enum": [
              "running",
              "stopped",
              "starting",
              "stopping",
              "error",
              "unknown"
            ],
            "title": "State"
          },
          "effective_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Port"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "allowed_ports": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Allowed Ports"
          },
          "managed": {
            "type": "boolean",
            "title": "Managed"
          }
        },
        "type": "object",
        "required": [
          "config",
          "generation",
          "observed_generation",
          "desired_state",
          "state",
          "effective_port",
          "started_at",
          "heartbeat_at",
          "last_error",
          "allowed_ports",
          "managed"
        ],
        "title": "MCPServiceView"
      },
      "ManagedMCPConfig": {
        "properties": {
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1024.0,
            "title": "Port",
            "default": 8081
          },
          "auto_start": {
            "type": "boolean",
            "title": "Auto Start",
            "default": true
          },
          "allowed_hosts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Hosts"
          },
          "allowed_origins": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Origins"
          },
          "request_timeout_s": {
            "type": "number",
            "maximum": 60.0,
            "exclusiveMinimum": 0.0,
            "title": "Request Timeout S",
            "default": 15
          },
          "max_request_bytes": {
            "type": "integer",
            "maximum": 1048576.0,
            "minimum": 1024.0,
            "title": "Max Request Bytes",
            "default": 65536
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ManagedMCPConfig"
      },
...
      "ModelResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "model_key": {
            "type": "string",
            "title": "Model Key"
          },
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "dimension": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimension"
          },
          "max_input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Input Tokens"
          },
          "tokenizer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokenizer Id"
          },
          "normalize": {
            "type": "boolean",
            "title": "Normalize"
          },
          "query_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Prefix"
          },
          "optimal_batch_size": {
            "type": "integer",
            "title": "Optimal Batch Size"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled"
          },
          "health_state": {
            "type": "string",
            "title": "Health State"
          },
          "checked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider_id",
          "display_name",
          "model_key",
          "capability",
          "dimension",
          "max_input_tokens",
          "tokenizer_id",
          "normalize",
          "query_prefix",
          "optimal_batch_size",
          "is_enabled",
          "health_state",
          "checked_at"
        ],
        "title": "ModelResponse"
      },
      "ModelTestResponse": {
        "properties": {
          "healthy": {
            "type": "boolean",
            "title": "Healthy"
          },
          "dimensions": {
            "type": "integer",
            "title": "Dimensions"
          },
          "tokens": {
            "type": "integer",
            "title": "Tokens"
          }
        },
        "type": "object",
        "required": [
          "healthy",
          "dimensions",
          "tokens"
        ],
        "title": "ModelTestResponse"
      },
...
      "ProviderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "family": {
            "type": "string",
            "title": "Family"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled"
          },
          "model_count": {
            "type": "integer",
            "title": "Model Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "family",
          "base_url",
          "is_enabled",
          "model_count"
        ],
        "title": "ProviderResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}